Restructued packages + added basic tests for model.
We were using the default package and that made it impossible to write unit tests in a separate folder.
I added all of the base files to a different package named `main`. Doing this allows us to use separate packages.

So far the only tests are for the model, but this works well because the model manages a lot of logic under-the-hood.

In future, the packaging should be changed so that only the top-level MVC stuff is in the `main` package, and the other
packages are put into their own sensibly-named packages.